Develop tests for sccp

+++ src/main/java/com/unifun/sigproxy/repository/sccp/SccpConcernedSignalingPointCodeConfigRepository.java
@@ -3,5 +3,5 @@ package com.unifun.sigproxy.repository.sccp;
 import com.unifun.sigproxy.models.config.sccp.SccpConcernedSignalingPointCodeConfig;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SccpConcernedSignalingPointCodeConfigRepository extends JpaRepository<SccpConcernedSignalingPointCodeConfig, Integer> {
+public interface SccpConcernedSignalingPointCodeConfigRepository extends JpaRepository<SccpConcernedSignalingPointCodeConfig, Long> {
 }
