+++ src/main/java/com/unifun/sigproxy/repository/sccp/SccpRemoteSignalingPointConfigRepository.java
@@ -3,5 +3,5 @@ package com.unifun.sigproxy.repository.sccp;
 import com.unifun.sigproxy.models.config.sccp.SccpRemoteSignalingPointConfig;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SccpRemoteSignalingPointConfigRepository extends JpaRepository<SccpRemoteSignalingPointConfig, Integer> {
+public interface SccpRemoteSignalingPointConfigRepository extends JpaRepository<SccpRemoteSignalingPointConfig, Long> {
 }
