+++ src/main/java/com/unifun/sigproxy/repository/sccp/SccpServiceAccessPointConfigRepository.java
@@ -3,5 +3,5 @@ package com.unifun.sigproxy.repository.sccp;
 import com.unifun.sigproxy.models.config.sccp.SccpServiceAccessPointConfig;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SccpServiceAccessPointConfigRepository extends JpaRepository<SccpServiceAccessPointConfig, Integer> {
+public interface SccpServiceAccessPointConfigRepository extends JpaRepository<SccpServiceAccessPointConfig, Long> {
 }
